+++ library/src/actions/partialCheck/index.ts
@@ -1,3 +1,3 @@
 export * from './partialCheck.ts';
 export * from './partialCheckAsync.ts';
-export * from './types.ts';
+export { PartialCheckIssue } from './types.ts';