+++ library/src/methods/forward/forward.ts
@@ -18,7 +18,7 @@ import type {
  * @returns The modified action.
  */
 export function forward<
-  TInput extends Record<string, unknown> | readonly unknown[],
+  TInput extends Record<string, unknown> | ArrayLike<unknown>,
   TIssue extends BaseIssue<unknown>,
 >(
   action: BaseValidation<TInput, TInput, TIssue>,