+++ library/src/methods/safeParser/safeParser.ts
@@ -62,9 +62,9 @@ export function safeParser(
     BaseSchema<unknown, unknown, BaseIssue<unknown>>,
     Config<BaseIssue<unknown>> | undefined
   > = (input: unknown) => safeParse(schema, input, config);
-  // @ts-ignore
+  // @ts-expect-error
   func.schema = schema;
-  // @ts-ignore
+  // @ts-expect-error
   func.config = config;
   return func;
 }